// packages/ProtobufViz/src/.test_data/gen/extra_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file extra.proto (package extra, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file extra.proto.
 */
export const file_extra: GenFile = /*@__PURE__*/
  fileDesc("CgtleHRyYS5wcm90bxIFZXh0cmEiFAoFRXh0cmESCwoDaW50GAEgASgFYgZwcm90bzM");

/**
 * @generated from message extra.Extra
 */
export type Extra = Message<"extra.Extra"> & {
  /**
   * @generated from field: int32 int = 1;
   */
  int: number;
};

/**
 * Describes the message extra.Extra.
 * Use `create(ExtraSchema)` to create a new message.
 */
export const ExtraSchema: GenMessage<Extra> = /*@__PURE__*/
  messageDesc(file_extra, 0);