// packages/ProtobufViz/src/.test_data/gen/spec_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file spec.proto (package spec, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Any } from "./google/protobuf/any_pb";
import { file_google_protobuf_any } from "./google/protobuf/any_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file spec.proto.
 */
export const file_spec: GenFile = /*@__PURE__*/
  fileDesc("CgpzcGVjLnByb3RvEgRzcGVjImYKA0ZvbxIMCgRuYW1lGAEgASgJEhYKA2JhchgCIAEoCzIJLnNwZWMuQmFyEhYKA2ZvbxgDIAEoCzIJLnNwZWMuRm9vEiEKA2FueRgEIAEoCzIULmdvb2dsZS5wcm90b2J1Zi5BbnkiHgoDQmFyEhcKBGZvb3MYASADKAsyCS5zcGVjLkZvb2IGcHJvdG8z", [file_google_protobuf_any]);

/**
 * @generated from message spec.Foo
 */
export type Foo = Message<"spec.Foo"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: spec.Bar bar = 2;
   */
  bar?: Bar;

  /**
   * @generated from field: spec.Foo foo = 3;
   */
  foo?: Foo;

  /**
   * @generated from field: google.protobuf.Any any = 4;
   */
  any?: Any;
};

/**
 * Describes the message spec.Foo.
 * Use `create(FooSchema)` to create a new message.
 */
export const FooSchema: GenMessage<Foo> = /*@__PURE__*/
  messageDesc(file_spec, 0);

/**
 * @generated from message spec.Bar
 */
export type Bar = Message<"spec.Bar"> & {
  /**
   * @generated from field: repeated spec.Foo foos = 1;
   */
  foos: Foo[];
};

/**
 * Describes the message spec.Bar.
 * Use `create(BarSchema)` to create a new message.
 */
export const BarSchema: GenMessage<Bar> = /*@__PURE__*/
  messageDesc(file_spec, 1);